requirements.txt
Keras==2.2.5
SQLAlchemy==1.3.18
pandas==0.25.1
numpy==1.18.0
scikit-learn==0.21.3
tensorflow-gpu==2.1.0
tqdm==4.41.1
python-dateutil
lightgbm==2.3.1
psycopg2-binary==2.8.4
hyperopt==0.2.3
hyperas==0.4.1
matplotlib==3.1.1
scipy==1.3.1
shap==0.35.0
xarray
pmdarima
xgboost